Exaltador/requirements.txt
fastapi
httpx
python-dotenv
pydantic
uvicorn
